Update Go 1.11 image to 1.11.6
Golang 1.11.4 contains an important fix for go modules checksum mismatches. Projects with symlinks or dependencies with symlinks produce different checksums in the go.sum file, which causes codebuild to fail if such a symlink exists and the checksum file is generated on a system other than linux.

This change bumps Golang and dep to the newest minor versions.